+++ HSTracker/Logging/CoreManager.swift
@@ -350,12 +350,12 @@ final class CoreManager: NSObject {
         Watchers.stop()
         MirrorHelper.destroy()
         let wm = game.windowManager
-        wm.constructedMulliganGuidePreLobby.viewModel.reset()
         if #available(macOS 10.15, *) {
             game.stopMulliganLivePolling()
             wm.rootOverlay?.viewModel.battlegroundsHeroPicking.reset()
             wm.rootOverlay?.viewModel.battlegroundsQuestPicking.reset()
             wm.rootOverlay?.viewModel.battlegroundsTrinketPicking.reset()
+            wm.rootOverlay?.viewModel.mulliganGuidePreLobby.viewModel.reset()
             wm.rootOverlay?.viewModel.mulliganGuide.reset()
             wm.rootOverlay?.viewModel.mulliganGuideV2.reset()
             wm.rootOverlay?.viewModel.constructedMulliganPreLobbyWidget.reset()