+++ HSTracker/UIs/Trackers/WindowManager.swift
@@ -65,10 +65,6 @@ class WindowManager {
         return $0
     }(MercenariesTaskListView(windowNibName: "MercenariesTaskListView"))
     
-    var constructedMulliganGuidePreLobby: ConstructedMulliganGuidePreLobby = {
-        return $0
-    }(ConstructedMulliganGuidePreLobby(windowNibName: "ConstructedMulliganGuidePreLobby"))
-    
     var flavorText: FlavorText = {
         return $0
     }(FlavorText(windowNibName: "FlavorText"))